removing warnings
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@5321 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/router/ROEdge.h
@@ -88,7 +88,7 @@ public:
 
     /** @brief Returns the number of edges this edge is connected to
         (size of the list of reachable edges) */
-    size_t getNoFollowing() const;
+    unsigned int getNoFollowing() const;
 
     /// Returns the edge at the given position from the list of reachable edges
     ROEdge *getFollower(size_t pos) const;
